pyproject.toml
[tool.poetry]
name = "fastapi_crud"
version = "0.1.0"
description = ""
authors = ["get-aakash <mail.aakash108@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
bandit = "^1.7.0"
fastapi = "^0.66.0"
flake8 = "^3.9.2"
pre-commit = "^2.13.0"
uvicorn = "^0.14.0"
SQLAlchemy = "^1.4.21"
black = "^21.6b0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
fastapi-mail = "^0.4.0"
python-dotenv = "^0.18.0"
PyJWT = "^2.1.0"
Jinja2 = "^3.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
